refactor: move `circuits` module testing to `tests.rs`
