+++ src/circuits/mod.rs
@@ -1,2 +1,3 @@
 pub mod merkle_sum_tree;
 pub mod utils;
+pub mod tests;
